requirements.txt
pytest~=7.0.1
Faker~=13.11.1
allure-pytest
pytest-xdist==2.1.0
requests==2.24.0
playwright~=1.26.1